+++ src/components/note/note.js
@@ -4,7 +4,7 @@ import themes from '../themes';
 
 class Note extends React.Component {
   render() {
-    const { theme, title, children, image } = this.props;
+    const { theme, title, children } = this.props;
     return (
       <div
         className="dr-ui--note py18 px18 round flex-parent flex-parent--row mb18"
@@ -13,11 +13,10 @@ class Note extends React.Component {
           color: themes[theme].styles.color
         }}
       >
-        {image && (
-          <div className="flex-child mr18 none block-mm pt3">
-            {themes[theme].image}
-          </div>
-        )}
+        <div className="flex-child mr18 none block-mm pt3">
+          {themes[theme].image}
+        </div>
+
         <div className="flex-child prose">
           <div className="txt-bold txt-m mb6">
             {title || themes[theme].label}
@@ -30,14 +29,12 @@ class Note extends React.Component {
 }
 
 Note.defaultProps = {
-  theme: 'default',
-  image: true
+  theme: 'default'
 };
 
 Note.propTypes = {
   title: PropTypes.string,
   children: PropTypes.node.isRequired,
-  image: PropTypes.bool,
   theme: PropTypes.oneOf([
     'warning',
     'error',